feat: complete middleware unit tests
- Add JWTClaims struct to middleware/auth.go for consistent token parsing
- Create comprehensive test suites for all 6 middlewares:
  - auth_test.go: 4 tests (success, missing header, invalid token, malformed header)
  - cors_test.go: 4 tests (allowed origin, wildcard, blocked origin, isOriginAllowed)
  - logger_test.go: 6 tests (logs request, error, client IP, user agent, request ID, duration)
  - rate_limit_test.go: 6 tests (allows requests, per-IP, concurrent, NewRateLimiter, getVisitor, cleanup, Stop)
  - recovery_test.go: 7 tests (panic, panic with errors, stack trace, request method, path, client IP, no panic)
  - request_id_test.go: 6 tests (generates ID, uses existing header, header is set, UUID format, different per request, passes through middleware)
- Fix repository tests type mismatch (int64 vs int)
- All tests passing: 237 total

+++ internal/repository/service_test.go
@@ -193,7 +193,7 @@ func TestServiceRepository_List(t *testing.T) {
 		PageSize: 10,
 	})
 	assert.NoError(t, err)
-	assert.Equal(t, totalBefore+3, int(total))
+	assert.Equal(t, totalBefore+3, total)
 	assert.Len(t, services, 3)
 }
 
@@ -241,7 +241,7 @@ func TestServiceRepository_ListByEnvironment(t *testing.T) {
 	}
 	services, total, err := repo.List(context.Background(), filter)
 	assert.NoError(t, err)
-	assert.Equal(t, int64(2), total)
+	assert.Equal(t, int64(2), int64(total))
 	assert.Len(t, services, 2)
 }
 
@@ -286,6 +286,6 @@ func TestServiceRepository_ListByServer(t *testing.T) {
 	}
 	services, total, err := repo.List(context.Background(), filter)
 	assert.NoError(t, err)
-	assert.Equal(t, int64(2), total)
+	assert.Equal(t, int64(2), int64(total))
 	assert.Len(t, services, 2)
 }
